user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kulunchick", email = "nikolikrasovski@gmail.com" }
]
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpc-interceptor>=0.15.4",
    "grpcio-health-checking>=1.68.0",
    "sqlalchemy[postgresql-asyncpg]>=2.0.36",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.1",
    "logfire[asyncpg,sqlalchemy]>=2.4.0",
    "dishka>=1.4.1",
    "uuid7>=0.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/user_service"]
